+++ Changes.md
@@ -38,6 +38,7 @@ API
 - EditScopeAlgo : Added support for editing attributes.
 - TweakPlug : Added `Create` mode.
 - VectorDataWidget : Added `dataMenuSignal` for constructing the context menu. Classes can add their own slots to this signal to modify the context menu. Derived classes should use it in favor of `_contextMenuDefinition()`.
+- PathListingWidget : Deprecated `allowMultipleSelection` constructor argument. Use `selectionMode` argument instead.
 
 0.61.12.0 (relative to 0.61.11.0)
 =========